Move more stuff to `utils`
