+++ tests/neural/genot_test.py
@@ -22,7 +22,7 @@ import optax
 
 from ott.geometry import costs
 from ott.neural.flow_models.models import VelocityField
-from ott.neural.flow_models.samplers import uniform_sampler
+from ott.neural.flow_models.utils import uniform_sampler
 from ott.solvers.linear import sinkhorn, sinkhorn_lr
 from ott.solvers.quadratic import gromov_wasserstein, gromov_wasserstein_lr
